create `make_export_df()` to prepare data frame for export for testing
Data frames stored as reactive values need to be exported before they can be tested by shinytest2. The x3p column in the data frames makes the shinytest2 snapshots 100+ MB, so drop this column before export.

Also, the source column stores the filepath for each x3p file, but the filepath is to the temp dir so it will change every time a test is run. Change the source column to only the filename before export.