change max for crosscut sliders to max y dimensions of the lands
Change the maximums for the crosscut sliders to the maximum y dimensions of the land x3ps in microns. The tricky part is that the rgl display needs the dimensions in the coordinate values in microns, and this is not necessarily the same as the number of rows in the x3p matrix.

The y maximums might not be integers, so the slider rounds down to the nearest integer to make the slider display cleaner.