rename processing helper fuctions with `_wrapper` at the end
For better clarity and consistency, I renamed the crosscut, grooves, signals, features, and scores helper functions as

- `get_aligned_signals_wrapper()`
- `get_bullet_scores_wrapper()`
- `get_ccdata_wrapper()`
- `get_default_cc_wrapper()`
- `get_features_wrapper()`
- `get_grooves_wrapper()`
- `get_signals_wrapper()`

I moved these functions to a single script called process-wrappers.R.

The terminology isn't great, but I will use "processing" to refer to the steps from getting the crosscuts to calculating the scores. The term "pre-processing" will refer to scaling, rotating, and anything else that needs to be done before finding the crosscuts.