renamed bullet-transformations.R as preprocess.R for consistency with process-wrappers.R
The functions in bullet-transformations.R are the preprocessing helpers that need to run before the wrapper functions in process-wrappers.R can be run, so I renamed bullet-transformations.R as preprocess.R for better clarity and consistency.

With the exception of `get_features_wrapper()`, the helper functions in process-wrappers.R are simply bulletxtrctr functions applied to each item in a data frame. This is why I added "wrapper" to the end of the function names. With the exception of `cond_x3p_m_to_mum()`, the functions in preprocess.R are not simple wrapper functions for bulletxtrctr functions, so I did not add "wrapper" to these functions names.