move report display to module
I want to avoid using modules as much as possible, at least for now, because they are harder to test and troubleshoot. However, the renderUI code for creating the report display takes up a ton of space in server.R. I added some helper functions, but this didn't reduce the space in server.R very much, so I moved the report display code to a module.

No reactive values are changed inside the module, and no values calculated inside the module are needed outside of it, so this makes me a little more comfortable using a module.