refactor (report-module): split report display into two parts
I'm still trying to figure out where in the pipeline the app freezes on large files. The renderUI in the report-module performs processing steps. As a first step in moving the processing steps out of the renderUI for better performance, I split the renderUI into two. The first renderUI now generates the top of the report (everything above the panels) and the second renderUI generates the collabsible panels. The first renderUI doesn't contain any processing code, but the second still does.

The app freezes somewhere after the "Rendering Report Objects" progress message. This message is displayed when the crosscut snapshots are created. As I write this, it just occured to me that maybe the creation of the snapshots themselves are crashing the app and not the subsequent code as I had previously assumed.