rename button ids to match labels
Looking at the server code, I constantly get confused as to which button ids belong to which buttons. So I changed the id of the file upload button from `bul_x3p` to `upload_button`, and the id of the "Add Bullet to Comparison List" button from `up_bull` to `add_to_list_button`.