fix: `downsample_bullet()` now outputs current bullet `cbull`
Previously, `downsample_bullet()` only output the data frame of all bullets `allbull`. It did not output the current bullet `cbull` even if it downsampled `cbull`. Now both `allbull` and `cbull` are output. If `allbull` or `cbull` is not downsampled, the function simply returns the input `allbull` or `cbull` without changes.