refactor: moved `read_bullet()` out of `preprocess_bullet()` for better clarity
I'm in the process of condensing the dataframes. I was reading through server.R, to find where `cbull` is created. I missed it at first because `cbull` is created inside `preprocess_bullet()`. Typically, the name of any function that I create that loads something from file starts with the word load. I moved the loading step out of `preprocess_bullet()`.