fix: increase downsampling of image previews to reduce memory (#35)
Images previews were previously downsampled by a factor of 5 (1 out of 5 pixels were kept), but the app would crash if the user tried to adjust the crosscuts on large x3p files (15+ MB). Image previews are now downsampled by a factor of 10. The app no longer crashes on large files.

It is important to note that only the image previews are downsampled, the underlying x3p data used in the statistical model is not changed.