fix: number lands sequentially
Previously, land names were assigned with the function `identify_lands()` look at the filenames for a bullet's lands. Starting at the beginning of the filenames, words that are present in all 6 filenames would be dropped. After finding a word that is not the same in all 6 filenames, the remaining parts of the filenames are returned as the land names. This resulted in unexpected land names for some files.

Now land names are simply numbered 1, 2, 3, 4, 5, 6. The numbers are stored as characters as before.