fix: fail with helpful error message if crosscut is NA
If `x3p_crosscut_optimize()` can't find a stable region in a land, it returns NA for the crosscut location. If the app continues, the crosscut slider will return an error message that the value cannot be NA. I need to talk with Heike to figure out proper treatment of lands without stable regions.

In the meantime, I added a check for NA values in `get_default_cc_wrapper()`. If any NA values are present, the app stops and returns a list of the bullet name and land name(s) that had NA values. The current statistical model cannot handle missing lands, so allowing users to proceed with the comparison could lead to wildly incorrect results.