docs: center screenshots in README
I added html formatting to the screenshots to center them on the page. For some reason that I can't figure out, this worked as expected for all screenshots except readme-run-app.png and readme-welcome.png. I had to manually change the markdown file itself instead of changing the Rmarkdown file and knitting the changes to create the markdown file.

The README displays R code but does not run any code, so we don't need the Rmarkdown. This commit deletes README.Rmd.